Fix email as login validation in personalize form (UserDataPanel).
This is for the case when email is used as login.  It checked that a
changed email address was valid as user id.  But the user id is never
changed here, only the login name.  We only need to
check if this address is not used by another user.

Also, we would get an error when changing an email address to for
example pete_o'hara+hello@example.org.  That one may look silly but is
a fine email adddress and should be allowed, also for login.  (Whether
this works everywhere is a different issue.)