Use schema tags to process ordering / filtering
Instead of munging the field manager after the schema (and any
FieldExtenders) have been processed, annotate the schema and
get autoform to do our munging for us.

As the omitting of fields has happened before a FieldExtender gets
to work, it can add fields without @@member-registration hiding them.

The downside is of course you can't order extended fields in
@@member-registration, but you can't please everyone.