Fix regression when visiting personal information as Manager.
Add extra test with Manager visiting both user and personal information.
In the initial backport of this fix, I saw a regression:
as Manager visit the user information of a member, then go to your personal information,
and you see the fullname, etc, of the member.
Tried on Plone 6 as well, but there it goes fine.